resource: moving the validation functions into validate
